# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Forex Bot API
  version: 0.1.0
  description: |
    REST API for the Forex/Crypto trading bot SaaS.

    **Owner:** Atlas Goro
    **Contract:** This spec is the source of truth. FastAPI runtime spec at `/openapi.json`
    MUST match — drift is enforced in CI.

    **Auth:** Bearer JWT (access token, 15 min TTL). Use `/auth/refresh` to rotate.

    **Errors:** All errors return the `ErrorResponse` shape. See `docs/api/error-contract.md`.
  contact:
    name: Backend (Atlas Goro)
    email: whyman404@gmail.com
  license:
    name: Proprietary
servers:
  - url: https://api.forexbot.app/api/v1
    description: Production
  - url: https://api.staging.forexbot.app/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local
security:
  - bearerAuth: []
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Create new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
            example:
              email: trader@example.com
              password: very-strong-passphrase-123
              display_name: Trader McTrade
      responses:
        '201':
          description: Account created, returns token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              email: trader@example.com
              password: very-strong-passphrase-123
              totp_code: '123456'
      responses:
        '200':
          description: Token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token, mint fresh access
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Confirm email with single-use token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyEmailRequest' }
      responses:
        '200':
          description: Verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request password reset email (always 200 to prevent enumeration)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: Email sent if user exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password using single-use token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '200':
          description: Reset done
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /auth/totp/enroll:
    post:
      tags: [auth]
      summary: Begin TOTP enrollment — returns secret + provisioning URI
      responses:
        '200':
          description: Pending — user must POST /auth/totp/verify to activate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TotpEnrollResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/totp/verify:
    post:
      tags: [auth]
      summary: Confirm TOTP enrollment with first valid code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpVerifyRequest' }
      responses:
        '200':
          description: TOTP activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /backtests:
    post:
      tags: [backtest]
      summary: Enqueue a backtest
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BacktestCreateRequest' }
            example:
              strategy_code: london_breakout_xauusd
              range_start: '2023-01-01'
              range_end: '2024-12-31'
              initial_balance: 10000.00
              params: { risk_per_trade_pct: 1.0 }
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacktestPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/PaymentRequired' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    get:
      tags: [backtest]
      summary: List backtests for current user
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BacktestPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /backtests/{backtestId}:
    parameters:
      - { name: backtestId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [backtest]
      summary: Get single backtest result
      responses:
        '200':
          description: Backtest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BacktestPublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /billing/checkout-session:
    post:
      tags: [billing]
      summary: Create a Stripe Checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutSessionRequest' }
      responses:
        '200':
          description: Session URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /billing/customer-portal:
    post:
      tags: [billing]
      summary: Generate Stripe Customer Portal URL
      responses:
        '200':
          description: Portal URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerPortalResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /billing/webhook:
    post:
      tags: [billing]
      summary: Stripe webhook receiver
      security: []
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Raw Stripe event payload }
      responses:
        '200':
          description: Processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /billing/me:
    get:
      tags: [billing]
      summary: Current subscription + recent invoices
      responses:
        '200':
          description: Subscription + invoices
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BillingMePublic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        15-minute access token from `/auth/login`. Rotate using `/auth/refresh`.
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    PaymentRequired:
      description: Active subscription required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until next attempt allowed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code
          example: AUTH_INVALID_CREDENTIALS
        message:
          type: string
          example: Invalid email or password.
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
          description: Correlate with logs and distributed traces
          example: 5f1f1c2a-3b22-4d04-90b7-d9a4e3a8f2e1
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
        display_name: { type: string, maxLength: 120, nullable: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        totp_code: { type: string, pattern: '^\d{6}$', nullable: true }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, default: Bearer }
        expires_in: { type: integer, description: Access token TTL (seconds) }
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    LogoutRequest:
      type: object
      properties:
        refresh_token: { type: string, nullable: true }
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token: { type: string }
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token: { type: string }
        new_password: { type: string, minLength: 12, maxLength: 128, format: password }
    TotpEnrollResponse:
      type: object
      required: [secret, provisioning_uri]
      properties:
        secret: { type: string }
        provisioning_uri: { type: string }
    TotpVerifyRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, pattern: '^\d{6}$' }
...
    BacktestCreateRequest:
      type: object
      required: [strategy_code, range_start, range_end, initial_balance]
      properties:
        strategy_code: { type: string }
        params: { type: object, additionalProperties: true }
        range_start: { type: string, format: date }
        range_end: { type: string, format: date }
        initial_balance: { type: number, exclusiveMinimum: 0, maximum: 10000000 }
    BacktestPublic:
      type: object
      required: [id, strategy_id, status, range_start, range_end, initial_balance, created_at]
      properties:
        id: { type: string, format: uuid }
        strategy_id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        range_start: { type: string, format: date }
        range_end: { type: string, format: date }
        initial_balance: { type: number }
        profit_factor: { type: number, nullable: true }
        max_drawdown_pct: { type: number, nullable: true }
        sharpe_ratio: { type: number, nullable: true }
        win_rate_pct: { type: number, nullable: true }
        total_trades: { type: integer, nullable: true }
        net_profit: { type: number, nullable: true }
        error_message: { type: string, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    CheckoutSessionRequest:
      type: object
      required: [plan_code]
      properties:
        plan_code: { type: string, enum: [trial, pro_monthly, pro_yearly, lifetime] }
        success_url: { type: string, format: uri, nullable: true }
        cancel_url: { type: string, format: uri, nullable: true }
    CheckoutSessionResponse:
      type: object
      required: [url, session_id]
      properties:
        url: { type: string, format: uri }
        session_id: { type: string }
    CustomerPortalResponse:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri }
    SubscriptionPublic:
      type: object
      properties:
        id: { type: string, format: uuid, nullable: true }
        plan:
          type: string
          enum: [free, trial, pro_monthly, pro_yearly, lifetime]
        status:
          type: string
          enum: [inactive, trialing, active, past_due, canceled, incomplete, unpaid]
        current_period_end: { type: string, format: date-time, nullable: true }
        cancel_at_period_end: { type: boolean, default: false }
        stripe_customer_id: { type: string, nullable: true }
    InvoicePublic:
      type: object
      required: [id, stripe_invoice_id, amount_cents, currency, status, created_at]
      properties:
        id: { type: string, format: uuid }
        stripe_invoice_id: { type: string }
        amount_cents: { type: integer, minimum: 0 }
        currency: { type: string }
        status: { type: string }
        paid_at: { type: string, format: date-time, nullable: true }
        hosted_invoice_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    BillingMePublic:
      type: object
      required: [subscription, invoices]
      properties:
        subscription: { $ref: '#/components/schemas/SubscriptionPublic' }
        invoices:
          type: array
          items: { $ref: '#/components/schemas/InvoicePublic' }
...